+++ changelog.md
@@ -12,6 +12,7 @@
 - Adds `pathlib.Path` support to `KeyStore` #148.
 - Adds Windows CI support #144.
 - Updates pyo3 to `0.20.0`
+- Updates time to `0.3.30`
 
 
 ## [0.14.1] - 2023-05-21