+++ changelog.md
@@ -11,8 +11,9 @@
 
 - Adds `pathlib.Path` support to `KeyStore` #148.
 - Adds Windows CI support #144.
-- Updates pyo3 to `0.20.0`
-- Updates time to `0.3.30`
+- Updates pyo3 to `0.20.0`.
+- Updates time to `0.3.30`.
+- Adds API for changing primary key expiry date #151.
 
 
 ## [0.14.1] - 2023-05-21