+++ changelog.md
@@ -15,6 +15,7 @@
 - Updates time to `0.3.30`.
 - Adds API for changing primary key expiry date via rjce #151.
 - Adds API for changing subkeys expiry date via rjce #152.
+- Updates sequoia dependency to 1.21.2.
 
 
 ## [0.14.1] - 2023-05-21