+++ changelog.md
@@ -16,6 +16,7 @@
 - Adds API for changing primary key expiry date via rjce #151.
 - Adds API for changing subkeys expiry date via rjce #152.
 - Updates sequoia dependency to 1.21.2.
+- Updates pyo3 dependency to 0.22.2.
 
 
 ## [0.14.1] - 2023-05-21